add integration tests that always fail
